+++ src/lapack/gpu/laset.cu
@@ -12,6 +12,7 @@
 
 #include <dlaf/gpu/assert.cu.h>
 #include <dlaf/lapack/gpu/laset.h>
+#include <dlaf/types.h>
 #include <dlaf/util_cublas.h>
 #include <dlaf/util_math.h>
